update bug fixed and readme, done deploy in gcp

+++ client/src/views/Dashboard.vue
@@ -6,13 +6,19 @@
         <template v-slot:activator="{ on: menu }">
           <v-tooltip bottom>
             <template v-slot:activator="{ on: tooltip }">
-              <v-btn color="primary" dark v-on="{ ...tooltip, ...menu }">Tag</v-btn>
+              <v-btn color="primary" dark v-on="{ ...tooltip, ...menu }"
+                >Tag</v-btn
+              >
             </template>
             <span>Search By Tag</span>
           </v-tooltip>
         </template>
         <v-list>
-          <v-list-item v-for="(item, index) in items" :key="index" @click="searchByTag(item.tag)">
+          <v-list-item
+            v-for="(item, index) in items"
+            :key="index"
+            @click="searchByTag(item.tag)"
+          >
             <v-list-item-title>{{ item.title }}</v-list-item-title>
           </v-list-item>
         </v-list>
@@ -51,7 +57,9 @@
 
 <script>
 // @ is an alias to /src
+import Swal from "sweetalert2";
 import AllArticle from "../components/AllArticle.vue";
+
 export default {
   name: "DashBoard",
   components: {
@@ -157,37 +165,58 @@ export default {
       }
     },
     deleteArticle(articleId) {
-      const token = localStorage.getItem("token");
-      this.axios({
-        method: "DELETE",
-        url: `/articles/${articleId}`,
-        headers: {
-          token
+      Swal.fire({
+        title: "Are you sure?",
+        text: "You will not be able to recover this imaginary file!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonText: "Yes, delete it!",
+        cancelButtonText: "No, keep it"
+      }).then(result => {
+        if (result.value) {
+          const token = localStorage.getItem("token");
+          this.axios({
+            method: "DELETE",
+            url: `/articles/${articleId}`,
+            headers: {
+              token
+            }
+          })
+            .then(({ data }) => {
+              let payloadAlert = {
+                text: data.message,
+                value: true
+              };
+              this.$emit("success-create", payloadAlert);
+            })
+            .catch(err => {
+              let text = "";
+              err.response.data.errors.forEach(element => {
+                text += element + ", ";
+              });
+              let payloadAlert = {
+                text,
+                value: true
+              };
+              this.$emit("error-create", payloadAlert);
+            });
+          Swal.fire(
+            "Deleted!",
+            "Your imaginary file has been deleted.",
+            "success"
+          );
+          // For more information about handling dismissals please visit
+          // https://sweetalert2.github.io/#handling-dismissals
+        } else if (result.dismiss === Swal.DismissReason.cancel) {
+          Swal.fire("Cancelled", "Your imaginary file is safe :)", "error");
         }
-      })
-        .then(({ data }) => {
-          let payloadAlert = {
-            text: data.message,
-            value: true
-          };
-          this.$emit("success-create", payloadAlert);
-        })
-        .catch(err => {
-          let text = "";
-          err.response.data.errors.forEach(element => {
-            text += element + ", ";
-          });
-          let payloadAlert = {
-            text,
-            value: true
-          };
-          this.$emit("error-create", payloadAlert);
-        });
+      });
+      //
     }
   },
   computed: {},
   created() {
-    this.fetchMyArticles();
+    this.fetchArticles;
   }
 };
 </script>
